requirements.txt
pymongo[srv]==3.12
python-dotenv
certifi
kagglehub
pandas